- build fix - youtube downloader update - add sync - youtube clip for live channel

+++ app/src/leanback/java/com/fongmi/android/tv/ui/adapter/ConfigAdapter.java
@@ -36,10 +36,15 @@ public class ConfigAdapter extends RecyclerView.Adapter<ConfigAdapter.ViewHolder
     }
 
     public int remove(Config item) {
+
+        int position = mItems.indexOf(item);
+        if (position == -1) return -1;
         item.delete();
-        mItems.remove(item);
-        notifyDataSetChanged();
+        mItems.remove(position);
+        notifyItemRemoved(position);
+
         return getItemCount();
+
     }
 
     @Override
